+++ RELEASE_NOTES.md
@@ -23,6 +23,9 @@
 
 - All library functionalities have been, for now, moved into the `internal` tree while further work
   is being done on the actual API and end-functionality of `gomod`.
+- `gomod` no longer wraps the invocation of `dot`. To get an image as output simply pipe `gomod`'s
+  output into the `dot` binary on the command-line. This means that the `--format` and `--visual`
+  flags have been removed and the the `--style` flag no longer implies `--visual`.
 
 ## 0.5.0